\name{adjustRestrictionsDF-package}
\alias{adjustRestrictionsDF-package}
\alias{adjustRestrictionsDF}
\docType{package}
\title{
\packageTitle{adjustRestrictionsDF}
}
\description{
\packageDescription{adjustRestrictionsDF}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{adjustRestrictionsDF}
\packageIndices{adjustRestrictionsDF}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
\packageAuthor{adjustRestrictionsDF}

Maintainer: \packageMaintainer{adjustRestrictionsDF}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in the R documentation ~~
~~ directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
